requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pandas
yfinance
ta
streamlit
plotly
requests
